+++ src/commonMain/kotlin/com/soywiz/korma/geom/triangle/Triangle.kt
@@ -53,11 +53,11 @@ interface Triangle {
     data class Base(override val p0: IPoint, override val p1: IPoint, override val p2: IPoint) : Triangle
 
     companion object {
-        private const val EPSILON: Float = 1e-12f
+        private const val EPSILON: Double = 1e-12
 
-        fun area(p1: IPoint, p2: IPoint, p3: IPoint): Float = area(p1.x, p1.y, p2.x, p2.y, p3.x, p3.y)
+        fun area(p1: IPoint, p2: IPoint, p3: IPoint): Double = area(p1.x, p1.y, p2.x, p2.y, p3.x, p3.y)
 
-        fun area(ax: Float, ay: Float, bx: Float, by: Float, cx: Float, cy: Float): Float {
+        fun area(ax: Double, ay: Double, bx: Double, by: Double, cx: Double, cy: Double): Double {
             val a = bx - ax
             val b = by - ay
             val c = cx - ax
@@ -213,7 +213,7 @@ fun Triangle.containsEdge(edge: Edge): Boolean = containsEdgePoints(edge.p, edge
 // In a triangle to check if contains and edge is enough to check if it contains the two vertices.
 fun Triangle.containsEdgePoints(p1: IPoint, p2: IPoint): Boolean = containsPoint(p1) && containsPoint(p2)
 
-private fun _product(p1: IPoint, p2: IPoint, p3: IPoint): Float = (p1.x - p3.x) * (p2.y - p3.y) - (p1.y - p3.y) * (p2.x - p3.x)
+private fun _product(p1: IPoint, p2: IPoint, p3: IPoint): Double = (p1.x - p3.x) * (p2.y - p3.y) - (p1.y - p3.y) * (p2.x - p3.x)
 
 fun Triangle.pointInsideTriangle(pp: IPoint): Boolean = if (_product(p0, p1, p2) >= 0) {
     (_product(p0, p1, pp) >= 0) && (_product(p1, p2, pp)) >= 0 && (_product(p2, p0, pp) >= 0)
@@ -271,7 +271,7 @@ fun Triangle.isPointAVertex(p: IPoint): Boolean = containsPoint(p)
 //for (var n:uint = 0; n < 3; n++) if (p == [this.points[n]]) return true;
 //return false;
 
-val Triangle.area: Float get() = Triangle.area(p0, p1, p2)
+val Triangle.area: Double get() = Triangle.area(p0, p1, p2)
 
 /** Alias for getPointIndexOffset */
 fun Triangle.index(p: IPoint): Int = this.getPointIndexOffsetNoThrow(p, 0, -1)
